move document type store into folder
