+++ src/backoffice/editors/document-type/actions/save/editor-action-document-type-save.element.ts
@@ -4,7 +4,7 @@ import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import type { UUIButtonState } from '@umbraco-ui/uui';
 import type { UmbNotificationDefaultData } from '../../../../../core/services/notification/layouts/default';
 import type { UmbNotificationService } from '../../../../../core/services/notification';
-import { UmbDocumentTypeStore } from '../../../../../core/stores/document-type.store';
+import { UmbDocumentTypeStore } from '../../../../../core/stores/document-type/document-type.store';
 import { UmbDocumentTypeContext } from '../../document-type.context';
 import { UmbContextConsumerMixin } from '@umbraco-cms/context-api';
